dpdp-backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.3.0
bcrypt==4.1.2
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cycler==0.12.1
ecdsa==0.19.0
fastapi==0.110.0
filelock==3.13.1
fonttools==4.50.0
fsspec==2024.3.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface_hub==0.23.0
idna==3.6
Jinja2==3.1.3
joblib==1.4.2
kiwisolver==1.4.5
lxml==5.2.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
mdurl==0.1.2
mpmath==1.3.0
networkx==3.4.2
numpy==1.26.4
packaging==24.0
pandas==2.2.2
passlib==1.7.4
pillow==10.3.0
pyasn1==0.5.1
pydantic==2.6.4
pydantic_core==2.16.3
Pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
PyYAML==6.0.1
regex==2024.5.15
reportlab==4.1.0
requests==2.31.0
rich==13.7.1
rsa==4.9
safetensors==0.4.3
scikit-learn==1.4.2
scipy==1.12.0
sentence-transformers==2.6.1
setuptools==69.5.1
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.29
starlette==0.36.3
sympy==1.12
threadpoolctl==3.4.0
tokenizers==0.19.1
torch==2.2.2
tqdm==4.66.2
transformers==4.40.1
typer==0.12.3
typing_extensions==4.12.0
urllib3==2.2.1
uvicorn==0.29.0
google-generativeai
google-generativeai
